# openapi.yaml
openapi: 3.0.0
info:
  title: Complex API Demo
  version: 1.0.0
  description: A comprehensive API demonstration for Unkey deployment showcase
  contact:
    name: Unkey Team
    url: https://unkey.dev
    email: support@unkey.dev
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /
    description: API server
...
paths:
...
  /api/analytics:
    get:
      summary: Get analytics data
      tags:
        - Analytics
      parameters:
        - in: query
          name: period
          schema:
            type: string
            enum:
              - last_7_days
              - last_30_days
              - last_90_days
              - last_year
            default: last_30_days
          description: Time period for analytics data
        - in: query
          name: granularity
          schema:
            type: string
            enum:
              - hour
              - day
              - week
              - month
            default: day
          description: Data granularity for time series
      responses:
        '200':
          description: Analytics data
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      period:
                        type: string
                        example: last_30_days
                      granularity:
                        type: string
                        example: day
                      summary:
                        type: object
                        properties:
                          totalUsers:
                            type: integer
                            example: 12500
                          activeUsers:
                            type: integer
                            example: 8750
                          newUsers:
                            type: integer
                            example: 2100
                          totalRevenue:
                            type: number
                            format: float
                            example: 156789.5
                          totalOrders:
                            type: integer
                            example: 3420
                          conversionRate:
                            type: number
                            format: float
                            example: 0.0275
                          avgOrderValue:
                            type: number
                            format: float
                            example: 45.85
                      timeSeries:
                        type: array
                        items:
                          type: object
                          properties:
                            timestamp:
                              type: string
                              format: date-time
                            users:
                              type: integer
                            revenue:
                              type: number
                              format: float
                            orders:
                              type: integer
                      topProducts:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                            name:
                              type: string
                            sales:
                              type: integer
                            revenue:
                              type: number
                              format: float
                      trafficSources:
                        type: object
                        properties:
                          organic:
                            type: object
                            properties:
                              users:
                                type: integer
                              percentage:
                                type: number
                                format: float
                          direct:
                            type: object
                            properties:
                              users:
                                type: integer
                              percentage:
                                type: number
                                format: float
                          social:
                            type: object
                            properties:
                              users:
                                type: integer
                              percentage:
                                type: number
                                format: float
                          referral:
                            type: object
                            properties:
                              users:
                                type: integer
                              percentage:
                                type: number
                                format: float
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Analytics data retrieved successfully
...
  /api/products:
    get:
      summary: Get paginated list of products
      tags:
        - Products
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          description: Number of items per page
        - in: query
          name: category
          schema:
            type: string
          description: Filter by product category
      responses:
        '200':
          description: Paginated list of products
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        name:
                          type: string
                        description:
                          type: string
                        price:
                          type: number
                          format: float
                        category:
                          type: string
                        inStock:
                          type: boolean
                        sku:
                          type: string
                        rating:
                          type: number
                          format: float
                        reviewCount:
                          type: integer
                        createdAt:
                          type: string
                          format: date-time
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      totalPages:
                        type: integer
                      hasNext:
                        type: boolean
                      hasPrev:
                        type: boolean
                  success:
                    type: boolean
                  message:
                    type: string
  /api/reports:
    get:
      summary: Generate and retrieve various types of reports
      tags:
        - Reports
      parameters:
        - in: query
          name: type
          required: true
          schema:
            type: string
            enum:
              - sales
              - users
              - products
              - analytics
              - orders
          description: Type of report to generate
        - in: query
          name: startDate
          schema:
            type: string
            format: date
          description: Start date for the report (YYYY-MM-DD)
        - in: query
          name: endDate
          schema:
            type: string
            format: date
          description: End date for the report (YYYY-MM-DD)
        - in: query
          name: format
          schema:
            type: string
            enum:
              - json
              - csv
              - pdf
            default: json
          description: Report output format
      responses:
        '200':
          description: Generated report data
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      reportId:
                        type: string
                        format: uuid
                      type:
                        type: string
                      period:
                        type: object
                        properties:
                          startDate:
                            type: string
                            format: date
                          endDate:
                            type: string
                            format: date
                      format:
                        type: string
                      generatedAt:
                        type: string
                        format: date-time
                      summary:
                        type: object
                      data:
                        type: array
                        items:
                          type: object
                      downloadUrl:
                        type: string
                        nullable: true
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
        '400':
          description: Bad request - missing or invalid parameters
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: null
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
...